Import 4.4BSD-Lite2

+++ sys/net/if_slvar.h
@@ -30,9 +30,9 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)if_slvar.h	8.3 (Berkeley) 2/1/94
+ *	@(#)if_slvar.h	8.4 (Berkeley) 1/9/95
  *
- * $Header: /cvsroot/src/sys/net/if_slvar.h,v 1.1.1.2 1998/03/01 02:10:07 fvdl Exp $
+ * $Header: /cvsroot/src/sys/net/if_slvar.h,v 1.1.1.3 1998/03/01 02:13:26 fvdl Exp $
  */
 
 /*
@@ -71,10 +71,10 @@ struct sl_softc {
 void	slattach __P((void));
 void	slclose __P((struct tty *));
 void	slinput __P((int, struct tty *));
-int	slioctl __P((struct ifnet *, int, caddr_t));
+int	slioctl __P((struct ifnet *, u_long, caddr_t));
 int	slopen __P((dev_t, struct tty *));
 int	sloutput __P((struct ifnet *,
 	    struct mbuf *, struct sockaddr *, struct rtentry *));
 void	slstart __P((struct tty *));
-int	sltioctl __P((struct tty *, int, caddr_t, int));
+int	sltioctl __P((struct tty *, u_long, caddr_t, int));
 #endif /* KERNEL */
