+++ sys/sys/disk.h
@@ -39,9 +39,9 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)disk.h	8.1 (Berkeley) 6/2/93
+ *	@(#)disk.h	8.2 (Berkeley) 1/9/95
  *
- * from: $Header: /cvsroot/src/sys/sys/disk.h,v 1.1.1.1 1998/03/01 02:11:02 fvdl Exp $ (LBL)
+ * from: $Header: /cvsroot/src/sys/sys/disk.h,v 1.1.1.2 1998/03/01 02:13:55 fvdl Exp $ (LBL)
  */
 
 /*
@@ -72,7 +72,7 @@ struct dkdriver {
 #ifdef notyet
 	int	(*d_open) __P((dev_t dev, int ifmt, int, struct proc *));
 	int	(*d_close) __P((dev_t dev, int, int ifmt, struct proc *));
-	int	(*d_ioctl) __P((dev_t dev, int cmd, caddr_t data, int fflag,
+	int	(*d_ioctl) __P((dev_t dev, u_long cmd, caddr_t data, int fflag,
 				struct proc *));
 	int	(*d_dump) __P((dev_t));
 	void	(*d_start) __P((struct buf *, daddr_t));
